kills overlays and idrs and prefabs
